+++ __tests__/SharedComponents.test.jsx
@@ -8,9 +8,9 @@ import React from 'react';
 import { renderWithRedux } from '../__utils__/render';
 
 import Banner from '../src/components/banner/banner';
-import ProgressBar from '../src/components/modal/progressbar';
+import ProgressBar from '../src/components/questionnaireModal/progressbar';
 import RedirectModal from '../src/components/about/redirectModal';
-import QuestionnaireModal from '../src/components/modal/questionnaireModal';
+import QuestionnaireModal from '../src/components/questionnaireModal/questionnaireModal';
 import ScrollIndicatorWrapper from '../src/components/scrollIndicatorWrapper/scrollIndicatorWrapper';
 import Spinner from '../src/components/spinner/spinner';
